backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-dotenv==1.1.0
pydantic==2.11.1
psycopg2-binary==2.9.10
pandas==2.2.3
plotly==6.1.0
openai==1.78.1
